Bird responds to mouse clicks/spacebar. Goes up and down. Movement/timing feels fairly similar to TB, can be adjusted in the Bird for now.
Gravity is currently on the bird instead of being a global that I can adjust (in case of gravity-enabled projectiles).
TODO: add global gravity, have individual objects able to override or scale said gravity for gameplay decisions.

InputManager added for a level of abstraction between raw inputs and actions

Added IInpuListener for any classes that might want to listen to InputManager events.