Learned that calling 'foreach' can cause garbage to accumulate, so switched them with simple for loops.
Also, newing the list of deleted bullets was not ideal, so instead keeping it around as a member variable.

This changes are meant to prevent the game from running out of memory after running it for about....let's say 3 years.

But which might be actually important on mobile devices as it's way easier to be literally out of memory there. Minimizing failure points.